fix clang warning
git-svn-id: https://svn.osgeo.org/gdal/trunk@40751 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/grib/degrib/degrib/myutil.c
@@ -757,7 +757,7 @@ void strTrimRight (char *str, char c)
       return;
    }
 
-   for (i = strlen (str) - 1;
+   for (i = (int)strlen (str) - 1;
         ((i >= 0) && (((unsigned char)isspace (str[i])) || (str[i] == c))); i--) {
    }
    str[i + 1] = '\0';
